# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IP2WHOIS LobeChat Plugin",
    "description": "A plugin for LobeChat that provides WHOIS information for IP addresses and domain names",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-deployment-url.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/whois": {
      "get": {
        "summary": "Get WHOIS information for an IP address or domain name",
        "operationId": "getWhoisInfo",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "description": "The IP address or domain name to lookup",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "The type of lookup to perform",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["ip", "domain"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IpWhoisResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DomainWhoisResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "IpWhoisResponse": {
        "type": "object",
        "properties": {
          "ip": {
            "type": "string",
            "description": "The IP address"
          },
          "country_code": {
            "type": "string",
            "description": "The country code"
          },
          "country_name": {
            "type": "string",
            "description": "The country name"
          },
          "region": {
            "type": "string",
            "description": "The region/state"
          },
          "city": {
            "type": "string",
            "description": "The city"
          },
          "postal_code": {
            "type": "string",
            "description": "The postal/zip code"
          },
          "latitude": {
            "type": "number",
            "description": "The latitude"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude"
          },
          "timezone": {
            "type": "string",
            "description": "The timezone"
          },
          "asn": {
            "type": "string",
            "description": "The Autonomous System Number"
          },
          "org": {
            "type": "string",
            "description": "The organization"
          },
          "isp": {
            "type": "string",
            "description": "The Internet Service Provider"
          },
          "network": {
            "type": "string",
            "description": "The network range"
          },
          "is_proxy": {
            "type": "boolean",
            "description": "Whether the IP is a proxy"
          },
          "is_tor": {
            "type": "boolean",
            "description": "Whether the IP is a TOR exit node"
          },
          "is_vpn": {
            "type": "boolean",
            "description": "Whether the IP is a VPN"
          }
        }
      },
      "DomainWhoisResponse": {
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "description": "The domain name"
          },
          "registrar": {
            "type": "string",
            "description": "The domain registrar"
          },
          "creation_date": {
            "type": "string",
            "format": "date-time",
            "description": "The domain creation date"
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "description": "The domain last updated date"
          },
          "expiration_date": {
            "type": "string",
            "format": "date-time",
            "description": "The domain expiration date"
          },
          "status": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "The domain status"
          },
          "nameservers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The domain name servers"
          },
          "dnssec": {
            "type": "string",
            "description": "The DNSSEC status"
          },
          "country_code": {
            "type": "string",
            "description": "The country code"
          },
          "country_name": {
            "type": "string",
            "description": "The country name"
          },
          "registrant": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "The registrant name"
              },
              "organization": {
                "type": "string",
                "description": "The registrant organization"
              },
              "country": {
                "type": "string",
                "description": "The registrant country"
              },
              "email": {
                "type": "string",
                "description": "The registrant email"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      }
    }
  }
}